+++ CHANGELOG.md
@@ -16,6 +16,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - do not configure it manually, always using Config UI X
   - required Homebridge v2.0.0 and above
 
+## [3.2.0] - (11.09.2024)
+
+## Changes
+
+- added possibility to set previous value for buttons [#154](https://github.com/grzegorz914/homebridge-melcloud-control/issues/154)
+- move some message to warning
+- presets previous settings apply improvement
+- cleanup
+
 ## [3.1.11] - (09.09.2024)
 
 ## Changes